+++ app/src/main/java/com/example/sandboxbank/App/core/deposit/domain/model/Deposit.kt
@@ -1,5 +1,7 @@
 package com.example.sandboxbank.App.core.deposit.domain.model
 
+import java.math.BigDecimal
+
 data class Deposit(
     override val id: Long,
     override val type: String,
@@ -7,6 +9,6 @@ data class Deposit(
     override val percentRate: Double,
     override val percentType: Long,
     override val period: Long,
-    override val balance: Long,
+    override val balance: BigDecimal,
     override val name: String = ""
 ) : FinancialItem()
