+++ app/src/main/java/com/example/sandboxbank/App/core/deposit/domain/model/FinancialItem.kt
@@ -1,5 +1,7 @@
 package com.example.sandboxbank.App.core.deposit.domain.model
 
+import java.math.BigDecimal
+
 sealed class FinancialItem {
     abstract val id: Long
     abstract val type: String
@@ -7,6 +9,6 @@ sealed class FinancialItem {
     abstract val percentRate: Double
     abstract val percentType: Long
     abstract val period: Long
-    abstract val balance: Long
+    abstract val balance: BigDecimal
     abstract val name: String
 }
